potential bugfix: removed showbar check in drawbar
Occasionally all windows would appear on the same workspace whenever a
window was quickly created and then destroyed (I think that was the
cause of the bug).
Looking at the commits, this check seems the most likely culprit.
Will now live with dwm without that check to see if the bug is
resolved